requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
python-docx==1.1.2
openai-whisper==20240930
libretranslatepy==2.1.4
anthropic==0.52.0
pydub==0.25.1
soundfile==0.13.1
gunicorn==23.0.0
